Fixed some spots where ushort was too small to hold the number of edges of a graph. Also removed the useless "completed if block" message as per Hayes's request.
Former-commit-id: f54982009e91e48775e7906d94dc599b9d4e8ebb

+++ src/Graph.cpp
@@ -430,7 +430,7 @@ void Graph::computeDistanceMatrix(vector<vector<short> >& dist) const {
 //Node ids should be between 0 and n-1
 void Graph::writeGraphEdgeListFormat(const string& fileName) {
     ushort n = getNumNodes();
-    ushort e = getNumEdges();
+    uint e = getNumEdges();
     ofstream outfile;
     outfile.open(fileName.c_str());
     outfile << n << " " << e << endl;
@@ -442,7 +442,7 @@ void Graph::writeGraphEdgeListFormat(const string& fileName) {
 
 void Graph::writeGraphEdgeListFormatPISWAP(const string& fileName) {
     ofstream outfile;
-    ushort e = getNumEdges();
+    uint e = getNumEdges();
     vector<string> nodeNames = getNodeNames();
     outfile.open(fileName.c_str());
     outfile << "INTERACTOR_A INTERACTOR_B" << endl;
